@@ -1,21 +1,19 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
   <Import Project="..\..\..\..\core-api.props" />
   <PropertyGroup>
-    
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <LangVersion>latest</LangVersion>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <DocumentationFile>bin\$(Configuration)\$(TargetFramework)\Identity.API.xml</DocumentationFile>
     <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
     <UserSecretsId>f7d51143-5f51-4aab-97ae-a4bff6b5ca7f</UserSecretsId>
     <DockerComposeProjectPath>..\..\..\..\docker-compose.dcproj</DockerComposeProjectPath>
-    <DocumentationFile>bin\$(Configuration)\$(TargetFramework)\Identity.API.xml</DocumentationFile>
     <DockerfileContext>..\..\..\..</DockerfileContext>
   </PropertyGroup>
 
   
  
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="5.1.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="3.1.*" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="3.1.*" />
     <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.0.2105168" />
   </ItemGroup>
 
@@ -28,6 +26,10 @@
     <ProjectReference Include="..\Identity.Mapping\Identity.Mapping.csproj" />
     <ProjectReference Include="..\Identity.Objects\Identity.Objects.csproj" />
     <ProjectReference Include="..\Identity.Validators\Identity.Validators.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Features\v1\Roles\" />
   </ItemGroup>
 
 </Project>
