@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Polly;
 using Polly.Retry;
@@ -21,46 +22,44 @@ namespace Identity.API
         public static async Task Main(string[] args)
         {
             IWebHost host = CreateWebHostBuilder(args).Build();
-            
-            using (IServiceScope scope = host.Services.CreateScope())
+
+            using IServiceScope scope = host.Services.CreateScope();
+            IServiceProvider services = scope.ServiceProvider;
+            ILogger<Program> logger = services.GetRequiredService<ILogger<Program>>();
+            IdentityContext context = services.GetRequiredService<IdentityContext>();
+            IHostEnvironment hostingEnvironment = services.GetRequiredService<IHostEnvironment>();
+            logger?.LogInformation("Starting {ApplicationContext}", hostingEnvironment.ApplicationName);
+
+            try
             {
-                IServiceProvider services = scope.ServiceProvider;
-                ILogger<Program> logger = services.GetRequiredService<ILogger<Program>>();
-                IdentityContext context = services.GetRequiredService<IdentityContext>();
-                IHostingEnvironment hostingEnvironment = services.GetRequiredService<IHostingEnvironment>();
-                logger?.LogInformation("Starting {ApplicationContext}", hostingEnvironment.ApplicationName);
-
-                try
+                if (!context.Database.IsInMemory())
                 {
-                    if (!context.Database.IsInMemory())
-                    {
-                        logger?.LogInformation("Upgrading {ApplicationContext}'s store", hostingEnvironment.ApplicationName);
-                        // Forces database migrations on startup
-                        RetryPolicy policy = Policy
-                            .Handle<SqlException>(sql => sql.Message.Like("*Login failed*", ignoreCase: true))
-                            .WaitAndRetryAsync(
-                                retryCount: 5,
-                                sleepDurationProvider: (retryAttempt => TimeSpan.FromSeconds(Math.Pow(2, retryAttempt))),
-                                onRetry: (exception, timeSpan, attempt, pollyContext) =>
-                                    logger?.LogError(exception, $"Error while upgrading database (Attempt {attempt})")
-                                );
-                        logger?.LogInformation("Starting {ApplicationContext} database migration", hostingEnvironment.ApplicationName);
+                    logger?.LogInformation("Upgrading {ApplicationContext}'s store", hostingEnvironment.ApplicationName);
+                    // Forces database migrations on startup
+                    RetryPolicy policy = Policy
+                        .Handle<SqlException>(sql => sql.Message.Like("*Login failed*", ignoreCase: true))
+                        .WaitAndRetryAsync(
+                            retryCount: 5,
+                            sleepDurationProvider: (retryAttempt => TimeSpan.FromSeconds(Math.Pow(2, retryAttempt))),
+                            onRetry: (exception, timeSpan, attempt, pollyContext) =>
+                                logger?.LogError(exception, $"Error while upgrading database (Attempt {attempt})")
+                            );
+                    logger?.LogInformation("Starting {ApplicationContext} database migration", hostingEnvironment.ApplicationName);
 
-                        // Forces datastore migration on startup
-                        await policy.ExecuteAsync(async () => await context.Database.MigrateAsync().ConfigureAwait(false))
-                            .ConfigureAwait(false);
-
-                        logger?.LogInformation($"Identity database updated");
-                    }
-                    await host.RunAsync()
+                    // Forces datastore migration on startup
+                    await policy.ExecuteAsync(async () => await context.Database.MigrateAsync().ConfigureAwait(false))
                         .ConfigureAwait(false);
 
-                    logger?.LogInformation($"Identity.API started");
+                    logger?.LogInformation($"Identity database updated");
                 }
-                catch (Exception ex)
-                {
-                    logger?.LogError(ex, "An error occurred on startup.");
-                }
+                await host.RunAsync()
+                    .ConfigureAwait(false);
+
+                logger?.LogInformation($"Identity.API started");
+            }
+            catch (Exception ex)
+            {
+                logger?.LogError(ex, "An error occurred on startup.");
             }
         }
 
