@@ -1,6 +1,9 @@
-﻿using MedEasy.DataStores.Core.Relational;
-using Identity.Objects;
+﻿using Identity.Objects;
+
+using MedEasy.DataStores.Core.Relational;
+
 using Microsoft.EntityFrameworkCore;
+using System;
 
 namespace Identity.DataStores.SqlServer
 {
@@ -14,7 +17,7 @@ namespace Identity.DataStores.SqlServer
         /// </summary>
         public DbSet<Account> Accounts { get; set; }
 
-        public DbSet<Claim> Claims { get; set; }
+        public DbSet<Role> Roles { get; set; }
 
         /// <summary>
         /// Builds a new <see cref="IdentityContext"/> instance.
@@ -27,12 +30,12 @@ namespace Identity.DataStores.SqlServer
         /// <summary>
         /// <see cref="DbContext.OnModelCreating(ModelBuilder)"/>
         /// </summary>
-        /// <param name="modelBuilder"></param>
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        /// <param name="builder"></param>
+        protected override void OnModelCreating(ModelBuilder builder)
         {
-            base.OnModelCreating(modelBuilder);
+            base.OnModelCreating(builder);
 
-            modelBuilder.Entity<Account>(entity =>
+            builder.Entity<Account>(entity =>
             {
                 entity.Property(x => x.Username)
                     .HasMaxLength(NormalTextLength)
@@ -49,54 +52,82 @@ namespace Identity.DataStores.SqlServer
                 entity.Property(x => x.PasswordHash)
                     .IsRequired();
 
-                entity.HasMany(x => x.Roles);
-
                 entity.Property(x => x.TenantId);
+
+                entity.Property(x => x.Email);
+
+                entity.Property(x => x.IsActive);
+
+                entity.Property(x => x.Locked);
+
+                entity.HasMany(x => x.Roles);
+                entity.HasMany(x => x.Claims)
+                      .WithOne()
+                      .HasForeignKey(accountClaim => accountClaim.AccountId)
+                      .HasPrincipalKey(account => account.Id);
             });
 
-            modelBuilder.Entity<Role>(entity =>
+            builder.Entity<AccountClaim>(entity =>
+            {
+                entity.HasKey(x => new { x.AccountId, x.Id });
+                entity.Property(x => x.Start);
+                entity.Property(x => x.End);
+                entity.OwnsOne(x => x.Claim, claim =>
+                {
+                    claim.Property(x => x.Type)
+                        .HasMaxLength(NormalTextLength)
+                        .IsRequired();
+
+                    claim.Property(x => x.Value)
+                        .HasMaxLength(NormalTextLength)
+                        .IsRequired();
+                });
+            });
+
+            builder.Entity<AccountRole>(entity => {
+                entity.HasKey(x => new { x.AccountId, x.RoleId });
+
+                entity.HasOne(ar => ar.Role)
+                      .WithMany(role => role.Accounts)
+                      .HasForeignKey(ar => ar.RoleId)
+                      .HasPrincipalKey(role => role.Id);
+
+                entity.HasOne(ar => ar.Account)
+                      .WithMany(account => account.Roles)
+                      .HasForeignKey(ar => ar.AccountId)
+                      .HasPrincipalKey(account => account.Id);
+            });
+
+            builder.Entity<RoleClaim>(entity =>
+            {
+                entity.HasKey(x => new { x.RoleId, x.Id });
+                entity.OwnsOne(x => x.Claim, claim =>
+                {
+                    claim.Property(x => x.Type)
+                        .HasMaxLength(NormalTextLength)
+                        .IsRequired();
+
+                    claim.Property(x => x.Value)
+                        .HasMaxLength(NormalTextLength)
+                        .IsRequired();
+                });
+            });
+
+
+            builder.Entity<Role>(entity =>
             {
                 entity.HasIndex(x => x.Code)
-                    .IsUnique();
+                   .IsUnique();
 
                 entity.Property(x => x.Code)
                     .HasMaxLength(ShortTextLength)
                     .IsRequired();
 
-                entity.HasMany(x => x.Claims);
-                entity.HasMany(x => x.Users);
-            });
-
-            modelBuilder.Entity<Claim>(entity =>
-            {
-                entity.HasIndex(x => x.Type)
-                    .IsUnique();
-
-                entity.Property(x => x.Type)
-                    .HasMaxLength(ShortTextLength)
-                    .IsRequired();
-                entity.Property(x => x.Value)
-                    .HasMaxLength(ShortTextLength)
-                    .IsRequired();
-
-                entity.HasMany(x => x.Roles);
-            });
-
-            modelBuilder.Entity<AccountClaim>(entity =>
-            {
-                entity.HasKey(x => new { x.AccountId, x.ClaimId });
-                entity.HasIndex(x => new
-                {
-                    x.AccountId,
-                    x.ClaimId,
-                    x.Start,
-                    x.End
-                }).IsUnique();
-            });
-
-            modelBuilder.Entity<RoleClaim>(entity =>
-            {
-                entity.HasKey(x => new { x.RoleId, x.ClaimId });
+                entity.HasMany(x => x.Claims)
+                      .WithOne()
+                      .HasForeignKey(roleClaim => roleClaim.RoleId)
+                      .HasPrincipalKey(role => role.Id);
+                entity.HasMany(x => x.Accounts);
             });
         }
     }
