@@ -1,6 +1,9 @@
 ﻿using MedEasy.Objects;
+using Optional;
+using Optional.Collections;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Identity.Objects
 {
@@ -11,12 +14,14 @@ namespace Identity.Objects
         /// <summary>
         /// Claims associated with the current <see cref="Role"/>
         /// </summary>
-        public IEnumerable<RoleClaim> Claims { get; set; }
+        public IEnumerable<RoleClaim> Claims => _claims;
+
+        private readonly IList<RoleClaim> _claims;
 
         /// <summary>
         /// <see cref="Account"/>s associated with the current role
         /// </summary>
-        public IEnumerable<Account> Users { get; set; }
+        public IEnumerable<AccountRole> Accounts { get; }
 
         /// <summary>
         /// Builds a new <see cref="Role"/> instance.
@@ -25,7 +30,64 @@ namespace Identity.Objects
         /// <param name="code"></param>
         public Role(Guid id, string code): base(id)
         {
-            Code = code;
+            Code = code ?? throw new ArgumentNullException(nameof(code));
+            _claims = new List<RoleClaim>();
+        }
+
+        /// <summary>
+        /// Adds or update a claim associated to the current role
+        /// </summary>
+        /// <param name="type"></param>
+        /// <param name="value"></param>
+        public void AddOrUpdateClaim(string type, string value)
+        {
+            if (type == null)
+            {
+                throw new ArgumentNullException(nameof(type));
+            }
+
+            if (value == null)
+            {
+                throw new ArgumentNullException(nameof(value));
+            }
+
+            Option<RoleClaim> optionRoleClaim = _claims.SingleOrNone(rc => rc.Claim.Type == type);
+
+            optionRoleClaim.Match(
+                some: rc => rc.Claim.ChangeValueTo(value),
+                none: () => _claims.Add(new RoleClaim(Id, Guid.NewGuid(), type, value))
+            );
+        }
+
+        /// <summary>
+        /// Removes all the claims with the specified type.
+        /// </summary>
+        /// <param name="type">type of claim to remove</param>
+        public void RemoveClaim(string type)
+        {
+            IEnumerable<RoleClaim> claimsToRemove = _claims.Where(rc => rc.Claim.Type == type)
+                                                           .ToArray();
+
+            foreach (RoleClaim rc in claimsToRemove)
+            {
+                _claims.Remove(rc);
+            }
+        }
+
+        /// <summary>
+        /// Removes only the claim with the specified <paramref name="type"/> and <paramref name="value"/>.
+        /// </summary>
+        /// <param name="type">type of claim to remove</param>
+        /// <param name="value">value of claim to remove</param>
+        public void RemoveClaim(string type, string value)
+        {
+            IEnumerable<RoleClaim> claimsToRemove = _claims.Where(rc => rc.Claim.Type == type && rc.Claim.Value == value)
+                                                           .ToArray();
+
+            foreach (RoleClaim rc in claimsToRemove)
+            {
+                _claims.Remove(rc);
+            }
         }
     }
 }
